refactor(ChannelFactory): refactor ChannelFactory for consistency in creation of dependencies
This commit adds some consistency to ChannelFactory in the creation of its dependencies. Previous to this commit, ChannelFactory both accessed the config directly and passed to the config for further processing. By standardizing around accessing the config in this factory alone, we can create a channel without additional ‘fromConfig’ factories. This allows the deletion of 2 factories, and more importantly, the config dependency is better contained within the codebase, leaving less parties dependent on its existence and contract.