feat: add ChannelProposalProcessorInterface and elevate client signing away from Channel
This commit moves the transactional duties of crafting a valid header out of Channel and into Client. It also takes steps to make parts of the lib more generic through removal of dependency on Chaincode types (see ProposalFactory). ChannelHeader only deals with things a Channel knows, and is annotated at Proposal time instead of the the opposite. Not sure if the interface stands the test of time, but it does provide the client with a centralized, repeatable place to build and process requests coming from a channel. It could be reduced into a “header signing” method, where consumers could assemble the proposal, but I’m not aware of a valid use case at current. We are missing tests around TransactionRequest but those are probably best handled in the upcoming refactor